// src/models/carListings/types.ts
export type TCarListing = {
  id: string;
  maker_id: string;
  model_id: string;
  year: number;
  mileage: number;
  price: number;
  condition: string;
  transmission: string;
  fuel_type: string;
  body_type: string;
  color: string;
  description: string;
  location: string;
  status: string;
};